+++ docs/changes.rst
@@ -6,6 +6,7 @@ Not yet released.
 * Fixed pull request message title.
 * Improved syntax error handling in Fluent format.
 * Fixed avatar display in notification e-mails.
+* Redirect /.well-known/change-password to /accounts/password/.
 
 `All changes in detail <https://github.com/WeblateOrg/weblate/milestone/82?closed=1>`__.